Allow float values for timeouts
(cherry picked from commit 6b37b45de1980b7fc64eb3206e41d86b8b886238)
(cherry picked from commit 3cf41879c0da7044f07b3a212f374f81494b18e1)
(cherry picked from commit ef490c828c868efd9c7c40e3c38867a704001a29)
(cherry picked from commit 5dc8341b8305510efef46bb4f030055b2f60a391)
(cherry picked from commit fde0617e5d989dd50b64442eee0feea66aef49e2)

+++ requests_retry_session/requests_retry_session.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2022-2025 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2022-2026 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -55,8 +55,8 @@ class RequestsRetryAdapterArgs(TypedDict, total=False):
     retries: int
     backoff_factor: float
     status_forcelist: Tuple[int, ...]
-    connect_timeout: int
-    read_timeout: int
+    connect_timeout: float
+    read_timeout: float
 
 
 def requests_session(adapter: requests.adapters.HTTPAdapter,
@@ -69,11 +69,12 @@ def requests_session(adapter: requests.adapters.HTTPAdapter,
     return session
 
 
-def requests_retry_adapter(retries: int = DEFAULT_RETRIES,
-                           backoff_factor: float = DEFAULT_BACKOFF_FACTOR,
-                           status_forcelist: Tuple[int, ...] = DEFAULT_STATUS_FORCELIST,
-                           connect_timeout: int = DEFAULT_CONNECT_TIMEOUT,
-                           read_timeout: int = DEFAULT_READ_TIMEOUT) -> TimeoutHTTPAdapter:
+def requests_retry_adapter(
+        retries: int = DEFAULT_RETRIES,
+        backoff_factor: float = DEFAULT_BACKOFF_FACTOR,
+        status_forcelist: Tuple[int, ...] = DEFAULT_STATUS_FORCELIST,
+        connect_timeout: float = DEFAULT_CONNECT_TIMEOUT,
+        read_timeout: float = DEFAULT_READ_TIMEOUT) -> TimeoutHTTPAdapter:
     retry = RetryWithLogs(
         total=retries,
         read=retries,
